refactor: Remove dead code and deduplicate content-locale utilities
- Delete unused locale-queries.ts, content-locale-context.ts + test
- Remove dead getCourseLocale export from course-tree-isolation.ts
- Generalize enforceSlugLocaleUniqueness + enforcePromptKeyLocaleUniqueness
  into single enforceFieldLocaleUniqueness(collection, field) factory
- Extract shared getNavItemsForLocale utility (Nav, MobileMenu, Footer)
- Extract shared getSystemLocale into src/i18n/server-locale.ts (layout, Footer)
