+++ src/server/payload/hooks/validateLocaleUniqueness.ts
@@ -1,5 +1,5 @@
 /**
- * Uniqueness hooks for (slug, locale) and (promptKey, locale) pairs
+ * Uniqueness hooks for (field, locale) pairs
  *
  * MongoDB compound uniqueness enforced via beforeChange hooks.
  */
@@ -7,13 +7,17 @@ import type { CollectionBeforeChangeHook } from 'payload'
 import { APIError } from 'payload'
 
 /**
- * Factory: enforces unique (slug, locale) per collection
+ * Factory: enforces unique (fieldName, locale) per collection.
+ * Works for slug+locale, promptKey+locale, or any other field.
  */
-export function enforceSlugLocaleUniqueness(collectionSlug: string): CollectionBeforeChangeHook {
+export function enforceFieldLocaleUniqueness(
+  collectionSlug: string,
+  fieldName: string = 'slug',
+): CollectionBeforeChangeHook {
   return async ({ data, req, operation, originalDoc }) => {
-    const slug = data?.slug ?? originalDoc?.slug
+    const fieldValue = data?.[fieldName] ?? originalDoc?.[fieldName]
     const locale = data?.locale ?? originalDoc?.locale
-    if (!slug || !locale) return data
+    if (!fieldValue || !locale) return data
 
     const selfId = operation === 'update' ? originalDoc?.id : undefined
 
@@ -21,7 +25,7 @@ export function enforceSlugLocaleUniqueness(collectionSlug: string): CollectionB
       collection: collectionSlug as any, // eslint-disable-line @typescript-eslint/no-explicit-any
       where: {
         and: [
-          { slug: { equals: slug } },
+          { [fieldName]: { equals: fieldValue } },
           { locale: { equals: locale } },
           ...(selfId ? [{ id: { not_equals: selfId } }] : []),
         ],
@@ -32,49 +36,11 @@ export function enforceSlugLocaleUniqueness(collectionSlug: string): CollectionB
 
     if (existing.docs.length > 0) {
       throw new APIError(
-        `A document with slug '${slug}' and locale '${locale}' already exists in ${collectionSlug}`,
+        `A document with ${fieldName} '${fieldValue}' and locale '${locale}' already exists in ${collectionSlug}`,
         400,
       )
     }
 
     return data
   }
-}
-
-/**
- * Enforces unique (promptKey, locale) on Prompts
- */
-export const enforcePromptKeyLocaleUniqueness: CollectionBeforeChangeHook = async ({
-  data,
-  req,
-  operation,
-  originalDoc,
-}) => {
-  const promptKey = data?.promptKey ?? originalDoc?.promptKey
-  const locale = data?.locale ?? originalDoc?.locale
-  if (!promptKey || !locale) return data
-
-  const selfId = operation === 'update' ? originalDoc?.id : undefined
-
-  const existing = await req.payload.find({
-    collection: 'prompts' as any, // eslint-disable-line @typescript-eslint/no-explicit-any
-    where: {
-      and: [
-        { promptKey: { equals: promptKey } },
-        { locale: { equals: locale } },
-        ...(selfId ? [{ id: { not_equals: selfId } }] : []),
-      ],
-    },
-    limit: 1,
-    overrideAccess: true,
-  })
-
-  if (existing.docs.length > 0) {
-    throw new APIError(
-      `A prompt with key '${promptKey}' and locale '${locale}' already exists`,
-      400,
-    )
-  }
-
-  return data
 }
